requirements.txt
asgiref==3.11.0
Django==5.2.8
django-filter==25.2
djangorestframework==3.16.1
pillow==12.0.0
psycopg2-binary==2.9.11
sqlparse==0.5.3
typing_extensions==4.15.0
